# Issue 4597: Optimization of String splits for Metrics