--- a/client/src/test/java/io/pravega/client/netty/impl/FlowHandlerTest.java
+++ b/client/src/test/java/io/pravega/client/netty/impl/FlowHandlerTest.java
@@ -375,5 +375,5 @@ public class FlowHandlerTest {
         @Override
         public void close() {
         }
-    };
+    }
 }