**Change log description**  
Uses single-character input in String.split() for efficiency reasons.

**Purpose of the change**  
Fixes #4597.

**What the code does**  
This PR attempts to use a variant of the `String.split()` method implementation, which is optimized for single-character split patterns. There are several occurrences in `NameUtils` class that can exploit this approach, but by some reason a reggex expression is used. This PR just replaces these occurrences, as some of them are on the hot write path.

**How to verify it**  
We have executed a micro benchmark to compare both approaches. Improvement in performance of `String.split()` is consistently between 2x and 4x:
```
    public static void main(String[] args) {
        String toSplit = "this/is/a/segment";
        long iterations = 1000000;
        long iniTime = System.nanoTime();
        for (long i = 0; i < iterations; i++) {
            toSplit.split("[/]");
        }
        double timeReggex = System.nanoTime() - iniTime;
        System.out.println("Time Reggex: " + timeReggex/1000000.0 + "ms");

        iniTime = System.nanoTime();
        for (long i = 0; i < iterations; i++) {
            toSplit.split("/");
        }
        double timeNoReggex = System.nanoTime() - iniTime;
        System.out.println("Time No Reggex: " + timeNoReggex/1000000.0 + "ms");
        System.out.println("Improvement: " + timeReggex/timeNoReggex);
    }
```
These are results from multiple executions:
```
Time Reggex: 747.486985ms
Time No Reggex: 208.371695ms
Improvement: 3.587276981165796
```
```
Time Reggex: 775.19952ms
Time No Reggex: 272.466208ms
Improvement: 2.845121696705964
```
```
Time Reggex: 795.027535ms
Time No Reggex: 399.01521ms
Improvement: 1.9924742593145761
```
Even with larger number of iterations (100M), the improvement is still consistent:
```
Time Reggex: 47721.324557ms
Time No Reggex: 17042.942773ms
Improvement: 2.80006365054524
```