@RaulGracia 2020-03-09
@kevinhan88 metrics are being published, I have seen them in the dashboard.